adding base activity card
